paxmon: more refactoring
